+++ tests/Poland/CorpusChristiTest.php
@@ -63,6 +63,6 @@ class CorpusChristiTest extends PolandBaseTestCase implements YasumiTestCaseInte
      */
     public function testHolidayType(): void
     {
-        $this->assertHolidayType(self::REGION, self::HOLIDAY, $this->generateRandomYear(), Holiday::TYPE_OTHER);
+        $this->assertHolidayType(self::REGION, self::HOLIDAY, $this->generateRandomYear(), Holiday::TYPE_OFFICIAL);
     }
 }